The chooser window gets margins, a framed list, and an aligned magnifier
Five visual fixes in one layout pass, all with existing PuiKit mechanisms:
the magnifier centers on the field's text line (align="center"), the field
fills its slot instead of stopping at 60 units, the page gains a 5px inner
margin, a slight gap separates the search row from the list, and the list
sits in a bordered frame. The console's Frame widget moves to the shared
keyhac/ui/frame.py so both windows draw the same border. A pixel-exact
spacer keeps the magnifier equidistant from the window edge and the field;
margins and spacer collapse on a character grid.
