+++ keyhac/ui/console.py
@@ -25,6 +25,7 @@ from puikit.widgets import Button, Checkbox, DropDown, Label, LayoutView, LogVie
 
 from keyhac.core import log
 from keyhac.core.keymap import _AUTHORING_WINDOW
+from keyhac.ui.frame import Frame
 
 logger = log.getLogger("Console")
 
@@ -52,23 +53,6 @@ _LEVELS = [
 ]
 
 _HEALTH_TICK_INTERVAL = 0.1
-
-_BORDER_STYLE = Style(fg=(120, 120, 132))
-
-
-class Frame(LayoutView):
-    """A LayoutView that draws a clear border line around its own extent.
-    draw_border() also clips the hosted content to the interior, so children
-    can fill up to the line without painting over it."""
-
-    def __init__(self, layout, margin_px: float = 6.0,
-                 line_style: Style = _BORDER_STYLE):
-        super().__init__(layout, margin_px=margin_px)
-        self.line_style = line_style
-
-    def draw(self, ctx) -> None:
-        ctx.draw_border(self.line_style)
-        super().draw(ctx)
 
 
 class ConsoleWindow:
